*: move to "etcdserver/api/membership"
