+++ etcdserver/api/cluster.go
@@ -15,7 +15,7 @@
 package api
 
 import (
-	"github.com/coreos/etcd/etcdserver/membership"
+	"github.com/coreos/etcd/etcdserver/api/membership"
 	"github.com/coreos/etcd/pkg/types"
 
 	"github.com/coreos/go-semver/semver"
