+++ etcdserver/api/etcdhttp/peer_test.go
@@ -25,7 +25,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/coreos/etcd/etcdserver/membership"
+	"github.com/coreos/etcd/etcdserver/api/membership"
 	"github.com/coreos/etcd/pkg/testutil"
 	"github.com/coreos/etcd/pkg/types"
 	"github.com/coreos/etcd/rafthttp"
