+++ etcdserver/api/v2v3/cluster.go
@@ -15,7 +15,7 @@
 package v2v3
 
 import (
-	"github.com/coreos/etcd/etcdserver/membership"
+	"github.com/coreos/etcd/etcdserver/api/membership"
 	"github.com/coreos/etcd/pkg/types"
 
 	"github.com/coreos/go-semver/semver"
