+++ etcdserver/cluster_util.go
@@ -22,7 +22,7 @@ import (
 	"sort"
 	"time"
 
-	"github.com/coreos/etcd/etcdserver/membership"
+	"github.com/coreos/etcd/etcdserver/api/membership"
 	"github.com/coreos/etcd/pkg/types"
 	"github.com/coreos/etcd/version"
